canbus: j1939: remove device tree setup of nodes
need to clean up and add error handling since this
is not handled in device tree. might need to look
at macros for definition like K_MUTEX_DEFINE
