--- dts/bindings/can/zephyr,j1939-node.yaml
@@ -1,102 +1,0 @@
-# Copyright (c) 2026 Deere & Company
-# SPDX-License-Identifier: Apache-2.0
-
-description: |
-  J1939 virtual node on a CAN bus.
-
-  Each instance is a child of a CAN controller node, mirroring how I2C
-  devices sit under an I2C controller.  The unit address / reg value is
-  the J1939 source address claimed by that node.
-
-  Example:
-
-    &can0 {
-        status = "okay";
-
-        j1939_node0: j1939-node@80 {
-            compatible = "zephyr,j1939-node";
-            reg = <0x80>;
-
-            /* J1939 NAME fields (SAE J1939-81) */
-            industry-group          = <2>;    /* 2 = Agricultural & Forestry */
-            arbitrary-address-capable;
-        };
-    };
-
-compatible: "zephyr,j1939-node"
-
-include: base.yaml
-
-on-bus: can
-
-properties:
-  reg:
-    required: true
-    description: |
-      J1939 source address claimed by this node (0x00–0xFD).
-      0xFE is the null address, 0xFF is the global broadcast address;
-      neither may be used as a claimed source address.
-      This value also becomes the node's unit address in the DTS.
-
-  id-number:
-    type: int
-    default: 0
-    description: |
-      21-bit identity number field in the J1939 NAME (SAE J1939-81 §5.4.2).
-      Typically derived from a hardware serial number.
-
-  manufacturer-code:
-    type: int
-    default: 0
-    description: |
-      11-bit manufacturer code assigned by SAE (J1939-81 Table B10).
-
-  ecu-instance:
-    type: int
-    default: 0
-    description: |
-      3-bit ECU instance within a function (0–7).
-
-  function-instance:
-    type: int
-    default: 0
-    description: |
-      5-bit function instance within a vehicle system (0–31).
-
-  function:
-    type: int
-    default: 0
-    description: |
-      8-bit function code (SAE J1939-81 Table B11).
-      0 means non-specific (general ECU).
-
-  vehicle-system:
-    type: int
-    default: 0
-    description: |
-      7-bit vehicle system code (SAE J1939-81 Table B12, column per industry group).
-      0 means non-specific system.
-
-  vehicle-system-instance:
-    type: int
-    default: 0
-    description: |
-      4-bit vehicle system instance (0–15).
-
-  industry-group:
-    type: int
-    default: 0
-    description: |
-      3-bit industry group (SAE J1939-81 Table B9):
-        0 - Global
-        1 - On-Highway Equipment
-        2 - Agricultural and Forestry Equipment
-        3 - Construction Equipment
-        4 - Marine
-        5 - Industrial / Process Control / Stationary
-
-  arbitrary-address-capable:
-    type: boolean
-    description: |
-      When present, this node is capable of negotiating an alternative
-      address if a contention occurs during address claim (SAE J1939-81 §5.2).
