+++ subsys/canbus/j1939/J1939_Cfg.h
@@ -286,7 +286,7 @@
 /***************************************************************************************************
  * Interface Layer
  **************************************************************************************************/
-#include <zephyr/canbus/j1939_dt.h>
+#include <zephyr/canbus/j1939.h>
 #define J1939_NUM_NODES (J1939_DT_NUM_NODES)
