+++ subsys/canbus/j1939/port/j1939_zephyr_glue.c
@@ -9,8 +9,22 @@
 
 #include <zephyr/canbus/j1939.h>
 #include <J1939Timer.h>
+#include "CriticalSectionManager.h"
 
 LOG_MODULE_REGISTER(j1939, CONFIG_CAN_LOG_LEVEL);
+
+// TODO: we could definitely reduce the scope of this lock
+K_MUTEX_DEFINE(j1939_lock);
+
+void CriticalSection_Lock(void)
+{
+    k_mutex_lock(&j1939_lock, K_FOREVER);
+}
+
+void CriticalSection_Unlock(void)
+{
+    k_mutex_unlock(&j1939_lock);
+}
 
 #if defined(CONFIG_J1939_LEGACY_AUTO_INIT)
